// sanity/types.ts
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type ProjectsQueryResult = Array<{
  _id: string;
  _updatedAt: string;
  name: string | null;
  completionDate: string | null;
  image: string | null;
  technologies: Array<{
    _id: string;
    _type: "technology";
    _createdAt: string;
    _updatedAt: string;
    _rev: string;
    name?: string;
    image?: {
      asset?: {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
      };
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      alt?: string;
      _type: "image";
    };
    url?: string;
    imageUrl: string | null;
  }> | null;
  slug: string | null;
  projectUrl: string | null;
  repositoryUrl: string | null;
  elevatorPitch: string | null;
}>;
// Variable: projectQuery
// Query:     *[_type=="project" && slug.current==$slug][0]{      _id,      _updatedAt,      name,      completionDate,      "image": image.asset->url,      "technologies": technologies[]->{        ...,        "imageUrl": image.asset->url,      },      "slug": slug.current,      projectUrl,      repositoryUrl,      elevatorPitch,      content    }  
export type ProjectQueryResult = {
  _id: string;
  _updatedAt: string;
  name: string | null;
  completionDate: string | null;
  image: string | null;
  technologies: Array<{
    _id: string;
    _type: "technology";
    _createdAt: string;
    _updatedAt: string;
    _rev: string;
    name?: string;
    image?: {
      asset?: {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
      };
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      alt?: string;
      _type: "image";
    };
    url?: string;
    imageUrl: string | null;
  }> | null;
  slug: string | null;
  projectUrl: string | null;
  repositoryUrl: string | null;
  elevatorPitch: string | null;
  content: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: "span";
      _key: string;
    }>;
    style?: "blockquote" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "normal";
    listItem?: "bullet" | "number";
    markDefs?: Array<{
      href?: string;
      _type: "link";
      _key: string;
    }>;
    level?: number;
    _type: "block";
    _key: string;
  }> | null;
} | null;
// Variable: contactsQuery
// Query:     *[_type=="contact"]{        _id,        _updatedAt,        username,        linkToProfile,        "favicon": {          ...favicon,          "url": favicon.asset->url,        }    }  
export type ContactsQueryResult = Array<{
  _id: string;
  _updatedAt: string;
  username: string | null;
  linkToProfile: string | null;
  favicon: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    _type: "image";
    url: string | null;
  };
}>;
// Variable: aboutQuery
// Query:     *[_type=="about" && name==$name][0]{      _id,      _updatedAt,      name,      introduction,      "image1": {        ...image1,        "url": image1.asset->url,       },      "image2": {        ...image2,        "url": image2.asset->url,       },      "image3": {        ...image3,        "url": image3.asset->url,       },    }  
export type AboutQueryResult = {
  _id: string;
  _updatedAt: string;
  name: string | null;
  introduction: string | null;
  image1: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    title?: string;
    description?: string;
    footer?: string;
    alt?: string;
    _type: "image";
    url: string | null;
  };
  image2: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    title?: string;
    description?: string;
    footer?: string;
    alt?: string;
    _type: "image";
    url: string | null;
  };
  image3: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    title?: string;
    description?: string;
    footer?: string;
    alt?: string;
    _type: "image";
    url: string | null;
  };
} | null;